% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBasePlaceTypes.R
\name{addHhldStructure}
\alias{addHhldStructure}
\title{Add structure to the households data}
\usage{
addHhldStructure(Hhld)
}
\arguments{
\item{Hhld}{A data.frame of households containing population by age
group.}
}
\value{
A data.frame of households with added variables
}
\description{
\code{addHhldStructure} adds detail structure to
the households.
}
\details{
The function takes a data.frame of households in a region and adds
variables required by HhldAllocationModel.
}
