% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateHouseholds.R
\name{adjustHousingSupply}
\alias{adjustHousingSupply}
\title{Adjusts supply of housing to match total demand for a housing type}
\usage{
adjustHousingSupply(TotUnitDiff, Units_Bz)
}
\arguments{
\item{TotUnitDiff}{An number identifying the total difference in units where
a positive number indicates more demand than supply and a negative number
indicates more supply than demand.}

\item{Units_Bz}{A named numeric vector identifying the number of dwelling
units of the type in each Bzone.}
}
\value{
A list having two components:
BalancedUnits_Bz A named numeric vector giving the number of units by Bzone
which matches total demand, and
AdjUnits_Bz A named numeric vector giving the amount of adjustment that was
made to the original number of units in each Bzone.
}
\description{
\code{adjustHousingSupply} adjusts housing supply in each Bzone so that the
total matches housing demand for a housing type
}
\details{
Although the function which assigns a housing type to each household closely
matches the proportions of housing by type, it does not exactly match the
proportions. Furthermore the dwelling unit input numbers may not exactly
match the total number of households. For these reasons, the supply of
dwelling units by type may not match the demand. This function adjusts the
supply so that it matches demand. The units of a type in each Bzone are
adjusted in proportion to the total number of units of that type.
}
