% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCarSvcCharacteristics.R
\name{AssignCarSvcCharacteristics}
\alias{AssignCarSvcCharacteristics}
\title{Main module function to assign car service level to each household.}
\usage{
AssignCarSvcCharacteristics(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignCarSvcCharacteristics} assigns the car service level to each
household based on the Bzone where the household is located.
}
\details{
This function assigns the car service level to each household based on the
Bzone where the household is located.
}
