% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDevTypes.R
\name{AssignDevTypes}
\alias{AssignDevTypes}
\title{Main module function to assign a development type to each household.}
\usage{
AssignDevTypes(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignDevTypes} assigns a development type to each household.
}
\details{
This function assigns a development type to each household based on the
household housing type and Bzone and input assumptions about the proportion
of housing units by housing type and Bzone that are urban.
}
