% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignLocTypes.R
\name{AssignLocTypes}
\alias{AssignLocTypes}
\title{Main module function to assign a location type to each household.}
\usage{
AssignLocTypes(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignLocTypes} assigns a location type to each household.
}
\details{
This function assigns a location type to each household based on the
household housing type and Bzone and input assumptions about the proportions
of housing units by housing type and Bzone that are urban or town.
}
