% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate4DMeasures.R
\name{Calculate4DMeasures}
\alias{Calculate4DMeasures}
\title{Main module function that calculates 4D measures for each Bzone.}
\usage{
Calculate4DMeasures(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{Calculate4DMeasures} calculates 4D measures for each Bzone.
}
\details{
This module calculates several 4D measures by Bzone including density,
diversity (i.e. mixing of land uses), design (i.e. multimodal network design),
and destination accessibility.
}
