% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFuturePlaceTypes.R
\name{CalculateFuturePlaceTypes}
\alias{CalculateFuturePlaceTypes}
\title{Main module function that assigns the place-types (Bzone) to each
household and firm.}
\usage{
CalculateFuturePlaceTypes(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{CalculateFuturePlaceTypes} assigns place-types to each household and firm.
}
\details{
This function allocates placetypes to households and firms. It randomly assigns
firms to the placetypes and uses a linear model built on features of a household
to assign a household to a placetype.
}
