% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateHouseholds.R
\name{chooseLocations}
\alias{chooseLocations}
\title{Assigns Bzones for set of households}
\usage{
chooseLocations(SortedHhIds_, Capacity_Bz, Weights_Bz)
}
\arguments{
\item{SortedHhIds_}{A character vector of household IDs sorted in descending
order of household income.}

\item{Capacity_Bz}{A numeric named vector of the number of dwelling units
in each Bzone named with the Bzone names.}

\item{Weights_Bz}{A numeric named vector of attraction weights by Bzone
named with the Bzone names.}
}
\value{
A named character vector of Bzones named with the corresponding
household IDs.
}
\description{
\code{chooseLocations} assigns Bzones to households based on number of
dwelling units in each Bzone and Bzone weights.
}
\details{
This function assigns Bzone locations to households for a specified housing
type. Households are assigned to Bzones in descending order of their
incomes (e.g. highest income household gets first choice). The probability
that a household is assigned a Bzone is a function of the number of
available housing units in each Bzone and attraction weights of each Bzone.
}
