% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHousing.R
\name{estimateHousingModel}
\alias{estimateHousingModel}
\title{Estimate housing choice model}
\usage{
estimateHousingModel(Data_df, StartTerms_)
}
\arguments{
\item{Data_df}{A data frame containing estimation data.}

\item{StartTerms_}{A character vector of the terms of the model to be
tested in the model.}
}
\value{
A list which has the following components:
Type: a string identifying the type of model ("binomial"),
Formula: a string representation of the model equation,
PrepFun: a function that prepares inputs to be applied in the binomial model,
OutFun: a function that transforms the result of applying the binomial model.
Summary: the summary of the binomial model estimation results.
}
\description{
\code{estimateHousingModel} estimates a binomial logit model for choosing
between single family and multifamily housing
}
\details{
This function estimates a binomial logit model for predicting housing choice
(single family or multifamily) as a function of the supply of housing of
these types and the demographic and income characteristics of the household.
}
