% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBasePlaceTypes.R,
%   R/CalculateFuturePlaceTypes.R
\docType{data}
\name{HhAllocationModelCoeff_df}
\alias{HhAllocationModelCoeff_df}
\title{Household Allocation Model}
\format{A data.frame having the following components
\describe{
  \item{CHID}{a numeric indicating a area type}
  \item{CHDESC}{a string indicating a area type (City, Rural, Suburban, Town, and Urban)}
  \item{VAR}{a string indicating the variable names (ASC, Singleton, Children, CoupleNoKids,
  OnlyElderly, HhIncome1000)}
  \item{COEFF}{a numeric value for coefficients}
}}
\source{
CalculateBasePlaceTypes.R script.

CalculateFuturePlaceTypes.R script.
}
\usage{
HhAllocationModelCoeff_df

HhAllocationModelCoeff_df
}
\description{
A data.frame containing coefficients of variables needed to
implement household allocation model

A data.frame containing coefficients of variables needed to
implement household allocation model
}
\keyword{datasets}
