% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHousing.R
\name{ipf}
\alias{ipf}
\title{Iterative proportional fitting function.}
\usage{
ipf(Seed_ar, MrgnVals_ls, MrgnDims_ls, RmseTarget = 1e-05, MaxIter = 100)
}
\arguments{
\item{Seed_ar}{A array of starting numbers for the calculation. The
dimensions of this array must be the same as the dimensions of the desired
result and with the dimensions of the margins used at control totals.}

\item{MrgnVals_ls}{A list of vectors, matrices, or arrays for margin sum
control totals. The list needs to provide controls for each margin of the
array. For example, in the case of it's use in this module, two matrices are
used as control totals to balance a 3-dimensional array of housing units by
Bzone, housing type, and household income quartile. One matrix has control
totals by Bzone and housing type. The other matrix has control totals by
housing type and income quartile. Note that if one or more margins share the
same dimension, the totals for the shared dimension must be the same.}

\item{MrgnDims_ls}{A list of vectors where each vector identifies the
dimensions of the 'Seed_ar' that a margin control corresponds to. For example
if the 'Seed_ar' has 3 dimensions and a margin control matrix corresponds to
the 2nd and 3rd dimensions of the 'Seed_ar' (i.e. the 1st dimension of the
control matrix corresponds to the 2nd dimension of the array and the 2nd
dimension of the control matrix corresponds to the 3rd dimension of the
array), then the entry for that control matrix would be c(2,3). The number
of components in 'MrgnDims_ls' and the order of those components must be
consistent with 'MrgnVals_ls'.}

\item{RmseTarget}{A scalar numeric value specifying the maximum root mean square
error between the margin control values and the corresponding array sums. The
default value is 1e-7.}

\item{MaxIter}{A scalar numeric value specifying the maximum number of
iterations for balancing the array over the specified dimensions.}
}
\value{
An array containing values that meet the margin controls.
}
\description{
\code{ipf} fits values in array to match margins.
}
\details{
This function uses an iterative proportional fitting algorithm to calculate
the values in an array to so that specified margin sums are matched. The
function was written specifically to meet the needs of balancing housing
units by type and household income quartile by Bzone with the housing demand
for households by Azone. The function was not written for general use and
so does not contain checks to assure that function arguments are proper.
}
