% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateHouseholds.R
\name{LocateHouseholds}
\alias{LocateHouseholds}
\title{Main module function to assign the Bzone for all households.}
\usage{
LocateHouseholds(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{LocateHouseholds} assigns households to Bzones.
}
\details{
This function assigns households to Bzones based on the housing type and
income of each household and on the housing supply and location weight of
each Bzone.
}
