% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateWorkerJobs.R
\name{LocateWorkerJobs}
\alias{LocateWorkerJobs}
\title{Main module function to assign workers to job locations.}
\usage{
LocateWorkerJobs(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{LocateHouseholds} assigns workers to job locations.
}
\details{
This function assigns workers to Bzone job locations. Workers are assigned
randomly to job locations consistent with the number of total jobs in the
location.
}
