% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateWorkerJobs.R
\docType{data}
\name{LocateWorkerJobsSpecifications}
\alias{LocateWorkerJobsSpecifications}
\title{Specifications list for LocateWorkerJobs module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{NewSetTable}{specifications for new table created by module}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
LocateWorkerJobs.R script.
}
\usage{
LocateWorkerJobsSpecifications
}
\description{
A list containing specifications for the LocateWorkerJobs module.
}
\keyword{datasets}
