% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBasePlaceTypes.R
\name{predictLogit}
\alias{predictLogit}
\title{Make prediction from a multinomial logit model}
\usage{
predictLogit(ModelData, ModelCoeffs)
}
\arguments{
\item{ModelData}{A data.frame model data containing all the necessary variables.}

\item{ModelCoeffs}{A data.frame of coefficients of the variables required by the model.}
}
\value{
An array of predictions
}
\description{
\code{predictLogit} function makes prediction given a set of data
and corresponding coefficients
}
\details{
The function takes a data.frame of model data and coefficient data
that are required in making prediction of the type of area or development
for a household
}
