% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateBasePlaceTypes.R
\name{runLogit}
\alias{runLogit}
\title{Run a logit model iteratively}
\usage{
runLogit(HhldPop = NULL, ModelData = NULL, ModelCoeffs = NULL,
  TargetPop = NULL, TargetGroups = NULL, MaxIter = 10,
  QuitCriteria = 4000)
}
\arguments{
\item{HhldPop}{A data.frame containing population by for each household by age groups.}

\item{ModelData}{A data.frame of model data containing all the necessary variables.}

\item{ModelCoeffs}{A data.frame of coefficients of the variables required by the model.}

\item{TargetPop}{An array of target population to be matched by the logit model.}

\item{TargetGroups}{An array of target groups that corresponds to the choices
available in the ModelCoeffs table.}

\item{MaxIter}{Maximum number of iteration to run the logit model (Default: 10).}

\item{QuitCriteria}{The quitting criteria for the logit model (Default: 4000).}
}
\value{
An array of predictions
}
\description{
\code{runLogit} function iteratively runs a logit model
to match a given target.
}
\details{
The function takes a data.frame of model data and coefficient data to run a
a logit model and
for a household
}
