% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculate4DMeasures.R
\name{sumValsInDist}
\alias{sumValsInDist}
\title{Sum values for block groups within specified distances}
\usage{
sumValsInDist(DistCutoff, DataToSum_, Lat_, Lng_)
}
\arguments{
\item{DistCutoff}{A numeric value in miles specifying the straight line
distance in miles to use as the distance threshold.}

\item{DataToSum_}{A numeric vector of the block group values to sum
corresponding to all Bzones.}

\item{Lat_}{A numeric vector of the latitudes of the block group centroids
in the same order as DataToSum_.}

\item{Lng_}{A numeric vector of the longitudes of the block group centroids
in the same order as DataToSum_.}
}
\value{
A numeric vector of the sums of the values in DataToSum_ for block
groups within the DistanceCutoff of each Bzone.
}
\description{
\code{sumValsInDist} sums specified values for block groups whose centroids
are located within a specified distance cutoff.
}
\details{
This function sums for each block group, a set of block group values of
all the block groups whose centroids are within the specified distance of
the the block group.
}
