% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDemandManagement.R
\docType{data}
\name{TdmModel_ls}
\alias{TdmModel_ls}
\title{Travel demand management models}
\format{An object of class \code{list} of length 2.}
\source{
AssignDemandManagement.R
}
\usage{
TdmModel_ls
}
\description{
A list of components used to predict the reduction in household DVMT as a
consequence of participation in individualized marketing programs and
employee commute options programs
}
\keyword{datasets}
