% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateEmployment.R
\name{adjustEmployment}
\alias{adjustEmployment}
\title{Adjusts Bzone employment to match workers}
\usage{
adjustEmployment(EmpTarget, Emp_Bz)
}
\arguments{
\item{EmpTarget}{An number identifying the total number of jobs so that each
worker has a job.}

\item{Emp_Bz}{A named numeric vector identifying the employment of the type
in each Bzone.}
}
\value{
A list having two components:
BalancedEmp_Bz A named numeric vector giving the employment by Bzone which
matches workers, and
AdjEmp_Bz A named numeric vector giving the amount of adjustment that was
made to the original employment in each Bzone.
}
\description{
\code{adjustEmployment} adjusts Bzone employment by type so that total
employment matches total workers.
}
\details{
The function adjusts the input values of employment by type and Bzone so that
the total amount of employment for the region equals the total number of
workers for the region.
}
