% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocateEmployment.R
\name{adjustEmployment}
\alias{adjustEmployment}
\title{Adjusts Bzone employment to match workers}
\usage{
adjustEmployment(EmpTarget, Emp_, Names = NULL)
}
\arguments{
\item{EmpTarget}{An number identifying the total number of jobs to be
matched.}

\item{Emp_}{A numeric vector identifying the employment in each Bzone.}

\item{Names}{A character vector identifying names corresponding to the
Emp_ argument vector if any.}
}
\value{
A integer vector of the number of jobs by Bzone which sums to the
total. The positions correspond to the positions of the input vector of jobs
by Bzone.
}
\description{
\code{adjustEmployment} adjusts Bzone employment by type so that total
employment matches total workers.
}
\details{
The function adjusts the input values of employment by type and Bzone so that
the total amount of employment for the region equals the total number of
workers for the region.
}
