% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignDemandManagement.R
\docType{data}
\name{TdmModel_ls}
\alias{TdmModel_ls}
\title{Travel demand management models}
\format{A list containing two vectors:
\describe{
 \item{PropDvmtReduce}{named vector of the proportional reduction in DVMT due to TDM program participation (names are ECO and IMP)}
 \item{PropMilesPerWkr}{named vector of the ratio of work tour DVMT per worker to total DVMT (names are household size 1 to 8)}
 }}
\source{
AssignDemandManagement.R
}
\usage{
TdmModel_ls
}
\description{
A list of components used to predict the reduction in household DVMT as a
consequence of participation in individualized marketing programs and
employee commute options programs
}
\keyword{datasets}
